Percentage of asymptomatic patients
Note that this data is from the public health Slovakia and contains a time lag, additionally with the current load there can be limitations in the level of detail in reporting. The 2 columns refer to: 
-  Asymptomatic_inclNULL = if we count patients with clinical form = NULL (not filled in) as asymptomatic
-  Asymptomatic_exclNULL = if we count patients with clinical form = NULL as symptomatic
When you go back in time you will find these two values converge as the investigators revisit cases and provide more details (when time allows).